[PINOT-3691] Handle controller failover correctly in the server. (#561)
If the server fails to commit because it reached a wrong controller, we retry after holding.

If the server gets not-leader, refresh the controller leader co-ordinates, making sure
that only one thread does the refresh.

+++ pinot-core/src/main/java/com/linkedin/pinot/core/data/manager/realtime/LLRealtimeSegmentDataManager.java
@@ -343,7 +343,7 @@ public class LLRealtimeSegmentDataManager extends SegmentDataManager {
                 if (success) {
                   _state = State.COMMITTED;
                 } else {
-                  // Back off and retry from holding state again.
+                  segmentLogger.info("Could not commit segment. Retrying after hold");
                   hold();
                 }
               }
@@ -414,7 +414,12 @@ public class LLRealtimeSegmentDataManager extends SegmentDataManager {
     File destSeg = makeSegmentDirPath();
     final String segTarFileName = destSeg.getAbsolutePath() + TarGzCompressionUtils.TAR_GZ_FILE_EXTENTION;
     try {
-      _protocolHandler.segmentCommit(_currentOffset, _segmentNameStr, new File(segTarFileName));
+      SegmentCompletionProtocol.Response response = _protocolHandler.segmentCommit(_currentOffset, _segmentNameStr,
+          new File(segTarFileName));
+      if (!response.getStatus().equals(SegmentCompletionProtocol.ControllerResponseStatus.COMMIT_SUCCESS)) {
+        segmentLogger.warn("Received controller response {}", response);
+        return false;
+      }
       FileUtils.deleteQuietly(new File(segTarFileName));
     } catch (FileNotFoundException e) {
       segmentLogger.error("Tar file {} not found", segTarFileName, e);
